new: Set up foundation for 3.0 transforms
The XSLT 3.0 version of the s9api transformer and executable are a lot
more complicated because there are a ton of options.  Trying to follow
the basic principle of clojure.java.io where all the conversion back and
forth is hidden, but you still have functions that return objects, so
sn/compiler returns an XsltCompiler, etc.

Don't set ErrorListener on compiler

Seems easier to set it on the Transformer if needed

Add tests and a basic stylesheet w/input

Still TODO: add the .sef files and verify that they were successfully
imported and used (by adding a stylesheet that uses them)

Remove ErrorListener impl

The default error listener is virtually identical, so it's not necessary
here.  Moved it to test as an example, and to test ErrorListner is
correctly set.