refactor: Rename/rewrite "chaining" functions
This also includes a new protocol I may or may not keep.  Refactored and
renamed the chaining functions because what they directly do is reduce a
set of XsltTransformers into a single destination object.  The
"reduction" version is for logging, and it seems similar to the core
"reduction" function (showing each step of the reduce), hence the name.